backend/requirements.txt
pytest==7.2.2
Flask==2.2.2
Werkzeug==2.2.2
psycopg2-binary==2.9.9
requests==2.31.0
jinja2==3.0.0